requirements.txt
afinn==0.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
contractions==0.1.73
cycler==0.12.1
datasets==2.19.1
dill==0.3.8
filelock==3.14.0
fonttools==4.52.4
frozenlist==1.4.1
fsspec==2024.3.1
gensim==4.3.2
grpcio==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
hyperframe==6.0.1
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
portalocker==2.8.2
protobuf==5.27.1
pyahocorasick==2.1.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.4
pydantic_core==2.18.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
qdrant-client==1.9.1
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
scikit-learn==1.5.0
scikit-surprise==1.1.4
scipy==1.12.0
sentence-transformers==3.0.1
setuptools==70.0.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
sympy==1.12.1
tbb==2021.12.0
textsearch==0.0.24
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
